Factorization more the Nodes and Edges
