+++ bluepysnap/network.py
@@ -18,8 +18,11 @@
 import abc
 
 import pandas as pd
+import numpy as np
 from cached_property import cached_property
 
+
+from bluepysnap import utils
 from bluepysnap.exceptions import BluepySnapError
 
 
@@ -30,9 +33,20 @@ class NetworkObject(abc.ABC):
         """Initialize the top level NetworkObjects accessor."""
         self._circuit = circuit
 
-    @abc.abstractmethod
-    def _get_populations(self):
+    def _get_populations(self, cls, config):
         """Collects the different NetworkObjectPopulation and returns them as a dict."""
+        res = {}
+        for file_config in config:
+            storage = cls(file_config, self._circuit)
+            for population in storage.population_names:  # pylint: disable=not-an-iterable
+                if population in res:
+                    raise BluepySnapError("Duplicated {} population: '{}'".format(self.__class__.__name__, population))
+                res[population] = storage.population(population)
+        return res
+
+    @abc.abstractmethod
+    def _collect_populations(self):
+        """Should specify the self._get_populations arguments."""
 
     @cached_property
     def _config(self):
@@ -41,7 +55,7 @@ class NetworkObject(abc.ABC):
     @cached_property
     def _populations(self):
         """Cached population dictionary."""
-        return self._get_populations()
+        return self._collect_populations()
 
     @cached_property
     def population_names(self):
@@ -104,10 +118,51 @@ class NetworkObject(abc.ABC):
         """Returns all the NetworkObject properties present inside the circuit."""
         return set(prop for pop in self.values() for prop in pop.property_names)
 
+    def _get_ids_from_pop(self, fun_to_apply, returned_ids_cls):
+        """Get CircuitIds of class 'returned_ids_cls' for all populations using 'fun_to_apply'.
+
+        Args:
+            fun_to_apply (function): A function that returns the list of IDs for each population
+                and the population containing these IDs.
+            returned_ids_cls (CircuitNodeIds/CircuitEdgeIds): the class for the CircuitIds.
+
+        Returns:
+            CircuitNodeIds/CircuitEdgeIds: containing the IDs and the populations.
+        """
+        str_type = "<U{}".format(max(len(pop) for pop in self.population_names))
+        ids = []
+        populations = []
+        for pop in self.values():
+            pop_ids, name_ids = fun_to_apply(pop)
+            pops = np.full_like(pop_ids, fill_value=name_ids, dtype=str_type)
+            ids.append(pop_ids)
+            populations.append(pops)
+        ids = np.concatenate(ids).astype(np.int64)
+        populations = np.concatenate(populations).astype(str_type)
+        return returned_ids_cls.from_arrays(populations, ids)
+
     @abc.abstractmethod
     def ids(self, *args, **kwargs):
         """Resolves the ids of the NetworkObject."""
 
     @abc.abstractmethod
-    def get(self, *args, **kwargs):
+    def get(self, group=None, properties=None):
         """Returns the properties of a the NetworkObject."""
+        ids = self.ids(group)
+        properties = utils.ensure_list(properties)
+
+        unknown_props = set(properties) - self.property_names
+        if unknown_props:
+            raise BluepySnapError("Unknown properties required: {}".format(unknown_props))
+
+        res = pd.DataFrame(index=ids.index, columns=properties)
+        for name, pop in self.items():
+            global_pop_ids = ids.filter_population(name)
+            pop_ids = global_pop_ids.get_ids()
+            pop_properties = set(properties) & pop.property_names
+            # indices from Population and get functions are different so I cannot
+            # use a dataframe equal directly and properties have different types so cannot use a
+            # multi dim numpy array
+            for prop in pop_properties:
+                res.loc[global_pop_ids.index, prop] = pop.get(pop_ids, prop).to_numpy()
+        return res.sort_index()
